requirements.txt
yt-dlp
fastapi
uvicorn[standard]
aiofiles 
assemblyai
transformers
torch
soundfile
pydantic-settings
openai
pydub 
ffmpeg-python
